+++ docs/changes.md
@@ -1,7 +1,8 @@
 # dev
 Release date: `2018-??-??`
 
-
+### Core
+- [NXDRIVE-825](https://jira.nuxeo.com/browse/NXDRIVE-825): Rely on the Python client for all Nuxeo API calls (**breaking change**)
 
 
 # 3.1.0