+++ docs/changes.md
@@ -4,6 +4,11 @@ Release date: `2018-??-??`
 ### Core
 - [NXDRIVE-825](https://jira.nuxeo.com/browse/NXDRIVE-825): Rely on the Python client for all Nuxeo API calls (**breaking change**)
 
+#### Minor changes
+- Jenkins:Added the `PYTEST_ADDOPTS` parameter to the Drive-tests job
+- Jenkins: Removed the `ENABLE_PROFILER` parameter from the Drive-tests job
+- Packaging: Added `nuxeo` 2.0.0
+
 
 # 3.1.0
 Release date: `2018-05-23`
@@ -75,7 +80,6 @@ Changes in command line arguments:
 - Jenkins: Update the Nuxeo snapshot to 10.2
 - Packaging: Fix symlink creation in `deploy.sh`
 - Packaging: Merged OS specific requirements-*.txt into one file
-- Packaging: Added `nuxeo` 2.0.0
 - Packaging: Added `pyaml` 17.12.1
 - Packaging: Added `requests` 2.18.4
 - Packaging: Removed `cffi`, will be installed with `xattr`